fix: keep the gesture handler import out of the Dialog's web bundle (OK-61140)
